+++ src/main/scala/org/podval/tools/backend/BackendProject.scala
@@ -2,5 +2,5 @@ package org.podval.tools.backend
 
 import org.gradle.api.Project
 
-abstract class BackendProject(val project: Project) extends Logging(project):
+abstract class BackendProject(project: Project) extends Logging(project):
   def apply(): Unit