+++ src/main/scala/org/podval/tools/backend/ProjectWithBackend.scala
@@ -4,6 +4,8 @@ import org.gradle.api.Project
 import org.podval.tools.build.ScalaBackend
 
 final class ProjectWithBackend(
-  val project: Project,
+  project: Project,
   val backend: ScalaBackend
+) extends WithProject(
+  project
 )